Fix SDK publishing (GEA-11855)
The apt package is named "python3-github", not "pygithub".
<https://packages.debian.org/bookworm/python3-github>

Removed the installation of pip since we were not actually using it. In
reality we _should_ be installing Python packages with pip, not apt, but
I don't want to change too much here yet.

Also pinned the Node.js image version to make things more deterministic.