pangea-node-sdk: refactor publish pipeline around Git tags (GEA-12973)
Publishing to the npm registry is now triggered via the pushing of Git
tags instead of requiring the use of the one `release` branch. This
enables flexibility in publishing prerelease versions outside of the
`release` branch and shaves one step off of the manual release process
(merging to `release`).

Instructions on how to publish a new version going forward have been
added to `CONTRIBUTING.md`.

This only affects pangea-node-sdk because I did not want to disrupt how
the other projects are being worked on at the moment. However, the
changes here are versatile enough that it would be very easy for the
other projects to "onboard".