pangea-node-sdk: raise timeout of Audit tests (PAN-16504)
Timeout was being set to 60s early on in the file, but then overwritten
to 20s later on. This affects all tests here.

Both of these values are less than the default polling timeout of 120s,
so this patch removes the duplicate `jest.setTimeout()` and sets the
overall test timeout to the polling timeout plus some buffer. This way
the polling timeout is hit before the test timeout.