react-mui-share-file-viewer: fix ESM import
With the addition of `preserveModules`, the generated ESM entry became
`dist/esm/src/index.js`, but in `package.json` we had the entry defined
as `dist/esm/index.js`. This causes ESM imports to fail in places like
@pangeacyber/webcomponents. This patch fixes the issue by disabling
`preserveModules`. To help prevent this from happening again, publint is
also added since it catches this exact issue.